+++ packages/sdk/scripts/web-worker-plugin.ts
