+++ packages/sdk/src/index.ts
@@ -5,6 +5,8 @@ export * from "./helpers/transfer"
 
 // Export machines
 export * from "./machines"
+export * from "./machines/luminadex"
+export * from "./machines/luminadex/types"
 export * from "./machines/wallet"
 
 // Re-Export xstate
@@ -14,9 +16,7 @@ export * from "xstate"
 export * from "o1js"
 
 // Export internal types
-// export * from "./machines/luminadex"
-// export * from "./machines/luminadex/types"
 
-// // Re-Export needed types
+// Re-Export needed types
 // export type * from "comlink"
 // export type * from "@aurowallet/mina-provider"
